+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Added IcdTimeZoneInfo, a very light implementation of System.TimeZoneInfo for the .NET Compact Framework
  - Added ThreadedWorkerQueue - a threadsafe way to enqueue items and have a worker thread process them one at a time
  - Added eDaysOfWeek flags enum
+ - Added support for reading the primitive type double to IcdXmlReader and XmlUtils
 
 ### Changed
  - Repeater changed to use configured callbacks instead of a dumb event